# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Productos",
        "description": "API para gestionar productos",
        "contact": {
            "name": "Equipo de desarrollo",
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/categories": {
            "get": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Obtener todas las categorías",
                "operationId": "ac9c4d978ea4e5b72915ca9728e5b8b1",
                "responses": {
                    "200": {
                        "description": "Listado de categorías"
                    }
                }
            },
            "post": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Crear una nueva categoría",
                "operationId": "cca49a87d9b0ebb9a98ad7c154b71737",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nueva Categoria"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripción de la nueva categoría"
                                    },
                                    "parent_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Categoría creada con éxito"
                    },
                    "400": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/api/categories/filter": {
            "post": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Filtrar categorías",
                "operationId": "889ab6492eed9bb1ff9644b99b7ef633",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Categoria1"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripción de la categoría"
                                    },
                                    "parent_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Categorías filtradas"
                    },
                    "400": {
                        "description": "Error de validación"
                    }
                }
            }
        },
...
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Obtener una categoría por su ID",
                "operationId": "09767bb2e47dde77dd1e2591428b0ae1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la categoría",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categoría obtenida con éxito"
                    },
                    "404": {
                        "description": "Categoría no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Actualizar una categoría por su ID",
                "operationId": "7dcf71887e585de7f881e9eb98d86126",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la categoría",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nueva Categoria"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripción actualizada de la categoría"
                                    },
                                    "parent_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Categoría actualizada con éxito"
                    },
                    "400": {
                        "description": "Error de validación"
                    },
                    "404": {
                        "description": "Categoría no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Eliminar una categoría por su ID",
                "operationId": "9b61b65823f4a21f182506b34ddf805c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la categoría",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categoría eliminada con éxito"
                    },
                    "404": {
                        "description": "Categoría no encontrada"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Actualizar parcialmente una categoría por su ID",
                "operationId": "86e6caed8bdc9857e1b991b70f58b23f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la categoría",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nuevo nombre de categoría"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripción actualizada de la categoría"
                                    },
                                    "parent_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Categoría actualizada con éxito"
                    },
                    "400": {
                        "description": "Error de validación"
                    },
                    "404": {
                        "description": "Categoría no encontrada"
                    }
                }
            }
        },
        "/api/categories/{id}/products": {
            "get": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Obtener todos los productos de una categoría por su ID",
                "operationId": "13e080ee10756a6884fd684f78866729",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la categoría",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Productos de la categoría obtenidos con éxito"
                    },
                    "404": {
                        "description": "Categoría no encontrada"
                    }
                }
            }
        },
        "/api/categories/{id}/tree": {
            "get": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Obtener una categoría con su estructura de árbol por su ID",
                "operationId": "c4f4ef3bc501469442b5feed1e4f907f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la categoría",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categoría con su estructura de árbol obtenida con éxito"
                    },
                    "404": {
                        "description": "Categoría no encontrada"
                    }
                }
            }
        },
...
    }
}